requirements.txt
requests
blessed
psutil
rich
cryptofuzz
mnemonic
pycryptodome
pywin32
